Implemented logger
This commit implements a logger package with various functions for
logging messages to different severity levels and outputs. The Init
function initializes four loggers with different severity levels and
sets up their respective output files. It returns an error if any of
the log files fail to be created. The Info, Warn, Error, and Fatal
functions write messages to their respective loggers at their
corresponding severity levels. The output function is used internally
to format and write log messages to the specified logger and also to
the console. If the logger is not initialized, a panic occurs.

Additionally, an exception has been added to the gitignore file for
*.log files.